requirements.txt
discord==2.3.2
mysql-connector-python==9.0.0
python-dotenv==1.0.1